Add Makefile targets to deploy/cleanup edpm-networkers
It is very similar to what "make edpm_deploy" is doing but this will
deploy "networker" nodes. In some cases it may be useful for development
purpose to deploy both compute node AND networker node and with this
change it will be easier to do and distinguish compute nodes from
networkers.

Additionally this patch renames devsetup scripts
gen-edpm-compute-node.sh and edpm-compute-cleanup.sh to more generic
names as those scripts are also used for both EDPM compute and networker
nodes now.