Correct order of default route on crc node
Starting with OCP 4.14, nmstate/network manager is beginning to create
the `ospbr` bridge default route as "static" which causes it to have a
higher priority over the external `ens3` interface.

This causes external connectivity to drop after the nncp resource is
applied and prevents Zuul from connecting to collect logs and fails the
job [1].

There has always been two default routes but it hasn't caused a problem
until OCP 4.14 where the priority changed.

[1] https://sf.hosted.upshift.rdu2.redhat.com/logs/51/449951/5/check/podified-multinode-tempest-internal-integration-rhel9-osp18/b7bf897/job-output.txt