Optional host macvlan for storage network
This patch adds an optional environmental variable called
`NETWORK_STORAGE_MACVLAN` that when set will create a macvlan on top of
the storage VLAN network to hold the IPs instead of the VLAN itself.

This allows host to/from pod network communication on the same host
using the storage VLAN.

This is necessary for example when using Cinder's LVM driver with the
iSCSI transport protocol and using the IP assigned to the pod's storage
network attachment.

The iscsi initiator (client) is run on the host, but it won't have
network access to the iSCSI target port that gets opened in the cinder
volume network namespace.

We don't default to always create the macvlan because we are also
introducing an alternative solution in the cinder-operator that will
work for now (until we move the iSCSI target creation to privsep).

This alternate solution requires the human operator to configure the
cinder backend to use the host's storage network IP directly (the
operator will make internally required changes to make that work).