Add script to assist on cloning repos
This patch adds a new clone-operato-repo script that assists on
cloning operator's code. It allows user to optionally provide a commit hash to
be switched out at the end.
The script includes an option, called CHECKOUT_FROM_OPENSTACK_REF,
that retrieves the operator's reference from openstack-operator go.mod
and checkout that specific commit hash found.
This functionality is useful when deploying the control-plane to guarantee
that CRs installed are in sync with operators installed by the
openstack-operator.
CHECKOUT_FROM_OPENSTACK_REF is disabled by default for now.