Add deployment of the neutron-sriov-nic-agent on standalone node
In order to test adoption of the neutron-sriov-nic-agent, at least with
some very basic tests, this patch adds deployment of the agent on the
Tripleo standalone node.
It can be disabled with config option but by default it is enabled so it
can be used in the adoption CI without further changes.

Related: #OSPRH-2433