Run standalone optionally with three Swift replicas
If standalone is running with a single replica, Swift replicators and
rsync won't be started and data migration is therefore not possible.

This patch runs Swift optionally with three separate disks and three
replicas, thus enabling replicators to run and test migration workloads.
To achieve this, three additional disks are created and attached to the
edpm nodes. The disk images are sparse, thus not using much space as
long as there is no data stored within.

The additional disks are not only useful for the standalone node itself,
but also for dataplane nodes to test with multiple disks.

The default is unchanged; this feature needs to be explicitly enabled.