adress bugs found when enabling tls-e
this change is a follow up to the naming change of the
network attachments. while testing the tls-e feature for the
edpm nodes i noticed that some fo the inventory fields were not
generated properly when the mixed case names were used.
to avoid that i have just normalised the names to lowercase.