Use higher metric than route added with crc
A default route was added to enable us to use
BM network as ctlplane for baremetal provisioning.
Afterwards, it was changed and it does not work
on crc with OCP 4.15. It's because that route has
a lower metric and always used.

Probably BM usecase is broken now, but we can verify
that and remove the route later in a followup.

```
192.168.122.0   192.168.122.1   255.255.255.0   UG    425    0        0 ospbr
192.168.122.0   0.0.0.0         255.255.255.0   U     426    0        0 ospbr
```
This changes to increase the metric so it's not used by
default.

jira: https://issues.redhat.com/browse/OSPRH-6307
Signed-off-by: rabi <ramishra@redhat.com>