Add ceph option for deployment with 1 compute
The ceph deployment for multinode tripleo job seems to fail when
creating an image, apparently because we currently only have one
compute. To have a functional cluster with 1 compute, we set an initial
ceph config with osd_pool_default_size=1.

This change also fixes the deployment so the ComputeHCI is actually
defined. To do that we needed to change the compute node hostname to
computehci and define the storage management network for the compute
node.