add compat code for using datasouces
to break a ci depenency we need to carfully sequence our usage
of dataSources.

pr #846 broke the ablity for
https://github.com/openstack-k8s-operators/dataplane-operator/pull/920
to merge in the dataplane operator by hardcoding a replacement of the
datasource in the nova dataplane service with only the custom config.

as a result when we move the requried configmaps to the dataSources
it overriedes them and resulting in ipam configution failing because
the ssh key is not present.

this change adapts install yamls to fully transfrom the dataplane service
into the new form that will be intoduced in pr#920

then we can remove this temporay mesure once that is merged vai
pr #845

Related: OSPRH-6407