Clone openstack-operator repo during prepare calls
When running install_yamls with CHECKOUT_FROM_OPENSTACK_REF enabled,
the clone-operator-repo script make sure that we clone
openstack-operator repo, since we need it to check for operators
references. But when running with this flag disabled, some make
targets fail to deploy since the openstack-operator is referenced
but not previously cloned.