Fix fake update and get dataplane_dep right when there are multiple.
When passed from the _make variable in the `ci-framework/role/update`
the `FAKE_UPDATE` was becoming `True` with a capital T, so in order to
prevent any problem with it value, we just run the block if it's
different from its default.

In some deployment one can have this:

```
openstackdataplanedeployment.dataplane.openstack.org/edpm-deployment-post-ceph
openstackdataplanedeployment.dataplane.openstack.org/edpm-deployment-pre-ceph
```

as output from `oc get openstackdataplanedeployment -o name` even
after a deployment - here, it's the HCI case.

So we take only the first one to make sure we create a valid
deployment file.

Increase default timeout to 600 as I've seen timeout error in some
deployment update.