Do not require networking for virt-customize
Add --no-network to the virt-customize call during edpm VM preparation
to stabilize the edpm_compute make target.

In some hosts (debian sid) and with certain libguestfs version (1.52.1)
the virt-customize call that prepared the edpm VM image is unstable. It
turns out that the virt-customize call we run does not require networing
setup so we can simply disable it. This also slightly speeds up running
it.