Adds bits to set domain and enable tlse for adoption multinode ci jobs
As part of [1] this aims to enable tls for the adoption multinode ci. We
have to move some of the overcloud config to j2 templates so we can
include the right values for cloud_domain (dns domain). It also adds
the external network for the tripleo controllers as without it the tls
deployment code fails. Finally this includes the startup of the freeipa
server as a container on the undercloud. Also see related patch at [2].

[1] https://issues.redhat.com/browse/OSPRH-8973
[2] https://review.rdoproject.org/r/c/rdo-jobs/+/54102