Deploy Rook operator
This patch allows to plug Rook to the existing k8s-operators development
workflow. With this patch is possible to deploy Rook operator, and it
can be used to attach an external Ceph cluster (already consumed by
OpenStack) to OpenShift.
The same code can be used to deploy an internal Ceph cluster, but is out
of scope for the current workflow.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>