nat64_router.sh: don't fail when libvirt domain is started
For what I see, virt-install already starts the VM; when I then detach
from it with `ctrl+]`, the script fails on `start` attempt, because the
domain is already active.

Perhaps this `start` is not needed at all? But at the very least, the
make target shouldn't fail because of it.