tripleo: split network config control flags
Add granular tripleo netconfig control flags

In local multi-cell setups, we use tripleo netconfig for cell stacks.
Extend it to the main stack as well.

In today multi-node CI we keep it disabled, so this
addition doesn't change the status quo.

Add flags to managing tripleo default routes and to attach ext net.
This is only used with tripleo dev setups, but may be extended to edpm
dev envs if needed.

Align the config download templates to use them consistently

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>