Devsetup: Remove dhcp entry even without VM
When using the `attach_default_interface_cleanup` target of devsetup on
a system where the CRC VM is no longer present it will fail to remove
the static DHCP entry from the libvirt network.

This patch makes the `interfaces-setup-cleanup.sh` script more robust as
it will now remove the static entry even when the VM is gone (using the
static IP address assigned).