NNCP Set MTU for all VLANs
If we don't set the VLANs MTUs in the NNCP we will always get a default
value and not a value based on the bridge like it should.

With this patch we allow setting the VLAN MTU specifically for the OCP
cluster.

This is necessary when the underlying interface is a tunnel and not a
physical NIC.

By default it will be set to the same value as the network mtu.