Ensure the openstacks crd exists in kuttl tests
Instead of just checking for any openstack.org crd, check explicitely
for openstacks.operator.openstack.org, and wait until that is available.
In some CI jobs, we run kuttl test for other operators before the
openstack-operators one. That means that the current check interprets
the fact that there are some openstack crds as being ready. Then when it
tries to get openstacks.operator.openstack.org, it fails because it's
not present yet.

Also remove the '|| true' at the end of the command, since that ignores
any error that might occurr.