Adding designateext example networking
In real world use, designate's unbound, bind9s and/or mDNS pods all need
to be accessible to "external" clients. designateext is used to simulate
such a network. Note that even though it is is configured similar to the
storage, internalapi, etc. networks, this is for testing and
demonstration purposes only.