Add infra dependency to mariadb kuttl
infra-operator now provides shared resources for most of the operators,
including Topology, that is referenced by the services' operators CRs to
inject a given topologySpreadConstraints or Affinity to their resulting
Pods or deployments.
This patch simply adds the "infra" dependency to mariadb, to make sure
that a Topology CRD is deployed and available for kuttl tests.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>