Support a CRC equivalent installation of SNO
This change adapts the IPv6 lab Single-node-Openshift script to support
being installed in an IPv4-only environment which is as similar as
possible to CRC. This allows development to occur against a SNO
environment when required without bringing up a whole IPv6 lab.

Not counting download time SNO installs in 37 minutes vs 7 for CRC, so
we can't simply switch to SNO in all scenarios, but making it easier to
develop with is desirable in some circumstances.

CRC actually supports microshift OKD variants too. This change also
adds a CRC_PRESET variable to allow switching CRC variants but this is
left undocumented because microshift is missing components such as OLM
which the development workflow requires.