Workaround random issues in nmstate-handler setup
In Downstream we randomly hitting issues where nmstate-handler
setup fails due to missing scc rule in 'nmstate-handler' cluster
role.

The scc rule creation is conditional[1], and if openshift-apiserver
is not running or is being re created the condition evaluates
to false and then nmstate-handler setup just stucks.

In order to workaround this adding a wait condition for
api server pod.

[1] https://github.com/nmstate/kubernetes-nmstate/pull/1113

Related-Issue: OSPCIX-554