Wait for all operator deployments in openstack_init
When creating the initialization resource we should also wait for
the service operator deployments to be available. Otherwise it might
happen that later automation steps get triggered before the operators
are operational. This was seen in a minor update test where patching
the openstackversion resource already was performed as in [1].

[1] https://github.com/openshift/release/blob/bf86ff46e1ef1d7baa05793c192f87ca17972efa/ci-operator/step-registry/openstack-k8s-operators/kuttl/openstack-k8s-operators-kuttl-commands.sh#L162

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>