Fix issues with metallb on OKD
Slightly different approach, just get the latest (v0.14.2) release,
patch what needs to be patched and and apply with -k for kustomize.
No need to wait for operatorhub.io or the deployment to fail before
patching.

The deployment for webhook-server wants to pull
quay.io/metallb/controller:main for the webhook-server. Which seems
to be a moving target, and recently was expecting to find an
additional CRD for service.BGPStatuses. So we pin the controller to
v0.14.5, which is what the other controller images are pinned to in
this release.