Fix glance kuttl target
In the previous PR #1085 we added openstack_crds as target in glance,
but it results in breaking the current tests because it seems to
conflicts with the common_prep task.
As Glance PR 787 shows, when openstack_crds is called, memcached is
deleted (it can be seen from the events), and glance is unable to
continue the CR deployment because memcached is not redeployed anymore.
While we can investigate further the current issue (that should not
definitely happen), this patch fixes the immediate problem and unblocks
the glance patches.

Related: https://issues.redhat.com/browse/OSPRH-19261
Signed-off-by: Francesco Pantano <fpantano@redhat.com>