Implement passing over an additional env file
Enable passing an additional heat environment file
to the standalone or tripleo deploy scripts.

Simplified a bit by alee.  We take the extra env file
defined by ADDITIONAL_ENV_FILE, and copy it to the standalone/tripleo
node to /tmp/additional_env_file.yaml.  This path is fixed.

We then look for this fixed path in the relevant scripts executed
on the node.